requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.6
django-crispy-forms==2.0
mysqlclient==2.1.1
sqlparse==0.4.3
tzdata==2022.7

django-ckeditor==6.5.1
django-filter==23.2